+++ bindgen/src/gen_go/compounds.rs
@@ -3,7 +3,10 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 use paste::paste;
-use uniffi_bindgen::{backend::Literal, interface::Type, ComponentInterface};
+use uniffi_bindgen::{
+    interface::{Literal, Type},
+    ComponentInterface,
+};
 
 use super::CodeType;